requirements.txt
# Core — Python 3.11+ (3.13 uses CrewAI 1.x; pin locally if you need 0.80.x)
crewai>=1.0.0
crewai-tools>=0.17.0

openai>=1.50.0
anthropic>=0.34.0

PyMuPDF>=1.24.0
python-docx>=1.1.0

sentence-transformers>=3.0.0
scikit-learn>=1.5.0
torch>=2.4.0

pydantic~=2.11.9
openpyxl>=3.1.0
pandas>=2.2.0

apify-client>=1.8.0
google-api-python-client>=2.140.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

streamlit>=1.39.0
slack-sdk>=3.33.0

python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=9.0.0
pytest>=8.0.0
